requirements-deploy.txt
# ==========================================================
# Jubu Backend — Deployment Requirements
# Cleaned from pip freeze: removed local-path installs,
# macOS-only packages (appnope, PyAudio etc.), dev tools.
# jubu_chat and jubu_datastore installed separately in Dockerfile.
# ==========================================================

# -- Core web framework --
fastapi==0.115.7
starlette==0.45.3
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2

# -- Database --
SQLAlchemy==2.0.37
alembic==1.14.1
psycopg2-binary==2.9.10
redis==6.4.0

# -- Auth --
bcrypt==4.3.0
passlib==1.7.4
PyJWT==2.10.1

# -- LiveKit --
livekit==1.0.17
livekit-api==1.0.7
livekit-protocol==1.0.8
websockets==15.0.1

# -- AI / LLM --
openai==1.75.0
google-generativeai==0.8.4

# -- Google Cloud --
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.35.0
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0

# -- ElevenLabs --
elevenlabs==1.58.1

# -- AssemblyAI --
assemblyai==0.40.2

# -- Audio processing --
soundfile==0.13.1
pydub==0.25.1
webrtcvad==2.0.10

# -- NLP --
spacy==3.8.4
thinc==8.3.4

# -- ML / Torch --
torch==2.2.2
transformers==4.49.0
tokenizers==0.21.0
safetensors==0.5.3
huggingface-hub==0.29.1

# -- HTTP clients --
aiohttp==3.11.11
httpx==0.28.1
requests==2.32.3

# -- Data --
numpy==2.2.3
pandas==2.2.3

# -- Utilities --
loguru==0.7.3
python-dateutil
PyYAML==6.0.2
Jinja2==3.1.5
rich==13.9.4
tqdm==4.67.1
regex==2024.11.6
